requirements.txt
playwright==1.19.0
requests==2.26.0
PyYAML==6.0
